Refactor horseController

+++ src/services/ownerService.ts
@@ -1,3 +1,4 @@
+import OwnerNotFoundError from '../errors/ownerNotFoundError';
 import Owner from '../models/owner';
 import OwnerRepository from '../repositories/ownerRepository';
 
@@ -19,10 +20,22 @@ export default class OwnerService {
     }
 
     static async deleteOwner(id: string) {
+        const owner = await OwnerRepository.getOwnerById(id);
+
+        if (!owner) {
+            throw new OwnerNotFoundError('Owner not found');
+        }
+
         await OwnerRepository.deleteOwner(id);
     }
 
     static async updateOwner(id: string, updatedOwnerData: Partial<Owner>): Promise<Owner> {
+        const owner = await OwnerRepository.getOwnerById(id);
+
+        if (!owner) {
+            throw new OwnerNotFoundError('Owner not found');
+        }
+        
         return await OwnerRepository.updateOwner(id, updatedOwnerData);
     }
 }
